Add aliases for merge conflicts (one of which requires full quoting)
Quoting at the git configuration parsing level, that is. Which is
because, I shit you not, the semicolon is a comment character in this
bastardized-INI format. And single-quoted strings are not considered
tokens. I can't even.

I've limited this quoting to the single shell word that is an argument
to -c, in the hopes that a sufficiently smart Git implementation doesn't
require another level of shell to invoke the command, but (in light of
the token thing) that's probably not going to happen. Maybe if I
double-quote *all* args that are currently single quoted, but then they
will look like an angry sandpaper of backslashes.